Convert encoded MAE tokens to correct type
This solves an issue when training MAE with half-precision.
In that case, we want to widen the type of the encoded tokens
to the same dtype as the mask token.

+++ docs/source/getting_started/benchmarks/imagenette_benchmark.py
@@ -756,7 +756,7 @@ class MAEModel(BenchmarkModule):
         x_masked = utils.repeat_token(
             self.mask_token, (batch_size, self.sequence_length)
         )
-        x_masked = utils.set_at_index(x_masked, idx_keep, x_decode)
+        x_masked = utils.set_at_index(x_masked, idx_keep, x_decode.type_as(x_masked))
 
         # decoder forward pass
         x_decoded = self.decoder.decode(x_masked)
